+++ SageTUI.LinuxContainerTests/Program.fs
@@ -1,0 +1,7 @@
+module SageTUI.LinuxContainerTests.Program
+
+open Expecto
+
+[<EntryPoint>]
+let main argv =
+  runTestsInAssemblyWithCLIArgs [] argv